shaders: Always use floating points for texture coordinates.

+++ res/gamedata/shaders/gl/iostructs/v_shadow_aref.h
@@ -4,15 +4,15 @@ out gl_PerVertex { vec4 gl_Position; };
 struct 	a2v
 {
 	float4	P	; // POSITION;		// Object-space position
- 	int2	tc0	; // TEXCOORD0;		// Texture coordinates
+ 	float2	tc0	; // TEXCOORD0;		// Texture coordinates
 };
 
 layout(location = NORMAL)		in float4	v_static_Nh		; // NORMAL;	// (nx,ny,nz,hemi occlusion)
 layout(location = TANGENT)		in float4	v_static_T		; // TANGENT;	// tangent
 layout(location = BINORMAL)		in float4	v_static_B		; // BINORMAL;	// binormal
-layout(location = TEXCOORD0)	in int2		v_static_tc		; // TEXCOORD0;	// (u,v)
+layout(location = TEXCOORD0)	in float2	v_static_tc		; // TEXCOORD0;	// (u,v)
 #ifdef	USE_LM_HEMI
-layout(location = TEXCOORD1)	in int2		v_static_lmh	; // TEXCOORD1;	// (lmu,lmv)
+layout(location = TEXCOORD1)	in float2	v_static_lmh	; // TEXCOORD1;	// (lmu,lmv)
 #endif
 layout(location = POSITION)		in float4	v_static_P		; // POSITION;	// (float,float,float,1)
